Changes suggested by JM and adopted tests
